FEATURE: Add link checker module (#1)
* TASK: Load entries from database

* TASK: Add check links command controller

* TASK: Use urls from settings and refactor code

* TASK: Implement crawling of links in backend module

* TASK: Use sitegeist/fusionlinkprototypes for links to content module

* TASK: Implement crawling of external URLs

* TASK: Implement email notification

* TASK: Improve Readme

* TASK: Fix typo

* TASK: Remove concurrency parameter and use setting only

* TASK: Allow MariaDb as well

* TASK: Use german error code names

Source: https://wiki.selfhtml.org/wiki/HTTP/Statuscodes

* TASK: Add composer dependencies

* BUGFIX: Fix indentation of configuration

* TASK: Remove all urls to crawl

* Add flash message if no domains could be found in sites

* Append workspace name

* Sort entries by source

* Strip /sites/site-identifier

* Crawl for phone numbers

* Validate assets as well

* Move property annotation around for better IDE support

* Remove unused variable assignment

* Split crawl commands

* Fix link renderer

* Simplify code

* Add note that domains also need to be added to site management

* TASK: Apply migration TYPO3.Form-20160601101500

Adjust "Settings.yaml" to use validationErrorTranslationPackage instead
of translationPackage

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Twitter.Bootstrap-20161124204912

Adjusts code to package renaming from "TYPO3.Twitter.Bootstrap" to
"Neos.Twitter.Bootstrap"

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Form-20161124205254

Adjusts code to package renaming from "TYPO3.Form" to "Neos.Form"

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Party-20161124225257

Adjusts code to package renaming from "TYPO3.Party" to "Neos.Party"

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Setup-20161124230842

Adjusts code to Setup Renaming

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Imagine-20161124231742

Adjusts code to Imagine Renaming

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Setup-20161125014759

Adjusts settings path from TYPO3.Setup to Neos.Setup to Setup Renaming

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.SwiftMailer-20161130105617

Adjusts code to package renaming from "TYPO3.SwiftMailer" to
"Neos.SwiftMailer".

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Seo-20170127154600

Migrate to new namespace

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Flow-20180415105700

Add scalar type hint to CacheAwareInterface implementations.

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Neos-20180907103800

Migrate RTE formatting configuration to new format

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Neos.Ui-20190319094900

Migrate additional RTE formatting configuration to new format

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Flow-20190425144900

Adjusts code to FlashMessageContainer renaming from "\Neos\Flow\Mvc" to
"\Neos\Flow\Mvc\FlashMessage".

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Flow-20190515215000

Adjust "Settings.yaml" to new PSR-3 logging settings (see
https://github.com/neos/flow-development-collection/pull/1574)

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Flow-20200813181400

Adjust "Settings.yaml" to new naming of settings (see
https://github.com/neos/flow-development-collection/pull/2051)

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Flow-20201003165200

Make default ValueObjects embedded=false

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Flow-20201109224100

Adjust DB migrations to Doctrine Migrations 3.0

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Flow-20201205172733

This migration does not actually change any code. It just displays a
warning if a PHP file still refers to the no longer existing
ComponentInterface

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Flow-20201207104500

Adjust code to deprecation of ComponentContext/ComponentParameters

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* Check if page is reachable in case parent page is disabled

* [TASK] Add page title in target column and link node to edit page

* [TASK] Improve afx logic

* [DOCS] Extend documentation with missing information

* [TASK] Put Flow annotation as second annotation that var is the first one

* [TASK] Remove unused constructor